requirements.txt
asgiref==3.3.4
autopep8==1.5.7
Django==3.2.4
djangorestframework==3.12.4
export==0.1.2
Faker==8.6.0
Pillow==8.2.0
pycodestyle==2.7.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
